+++ src/backend/main/settings.py
@@ -179,6 +179,9 @@ class Base(Configuration):
     RESANA_AUTHSERVICE_ENDPOINT = values.Value(
         environ_name="RESANA_AUTHSERVICE_ENDPOINT", environ_prefix=None
     )
+    RESANA_OTP_CHALLENGE_TTL = values.IntegerValue(
+        120, environ_name="RESANA_OTP_CHALLENGE_TTL", environ_prefix=None
+    )
     RESANA_S3_BUCKET = values.Value(
         environ_name="RESANA_S3_BUCKET", environ_prefix=None
     )